- major refactoring: - handles relations (here: multipolygons) in a much more sophisticated way, tries to merge members of relations to valid polygons - computes roles of members in multipolygons and does not rely on given roles (they are occasionally invalid) - clipping of polygons and polygons with holes at tile boundaries - clipping of ways at tile boundaries - way simplification with Douglas-Peucker algorithm - refactored TileBasedDataStore classes - fixed bugs in HDTileBasedDataStore - cleaned up the code, removed warnings
git-svn-id: http://mapsforge.googlecode.com/svn/trunk@1563 a424f762-2dd8-11df-bfeb-7b4c664f286a

+++ mapsforge/src/org/mapsforge/preprocessing/map/osmosis/DeltaEncoder.java
@@ -59,8 +59,8 @@ class DeltaEncoder {
 		while (it.hasNext()) {
 			Integer currentLat = it.next();
 			Integer currentLon = it.next();
-			result.add((currentLat - prevLat));
-			result.add(currentLon - prevLon);
+			result.add(Integer.valueOf((currentLat.intValue() - prevLat.intValue())));
+			result.add(Integer.valueOf(currentLon.intValue() - prevLon.intValue()));
 
 			prevLat = currentLat;
 			prevLon = currentLon;
